UPDATE separate sidebar into separate parts
