+++ app/blog/design-system/buttons/page.tsx
@@ -6,7 +6,7 @@ import Post from '@/components/post'
 import PostBanner from '@/components/post/post.banner'
 import PostNote from '@/components/post/post.note'
 import PostSection from '@/components/post/post.section'
-import PostSideBar from '@/components/post/post.sidebar'
+import PostSideBar from '@/components/post/sidebar'
 import type { Metadata } from 'next'
 
 export const metadata: Metadata = { title: 'Buttons · Design System' }
